Refactored and Commented the Experts Search Code
This commit adds comments to most of the code used for experts searching along with some minor refactoring as needed for clarity.
