--- api/Doyen.API/Doyen.API/Elasticsearch/Settings/ElasticsearchSettings.cs
@@ -1,13 +1,0 @@
-﻿namespace Doyen.API.Elasticsearch.Settings
-{
-    public class ElasticsearchSettings : IElasticsearchSettings
-    {
-        public string Url { get; set; }
-
-        public string Username { get; set; }
-
-        public string Password { get; set; }
-
-        public int SearchRecordsLimit { get; set; }
-    }
-}
