+++ api/Doyen.API/Doyen.API/Experts/IExpertSearcher.cs
@@ -1,14 +1,34 @@
-﻿using Doyen.API.Dtos;
-using Doyen.API.Models;
+﻿using Doyen.API.Dtos.Requests;
+using Doyen.API.Dtos.Responses;
+using System.Collections.Generic;
 
 namespace Doyen.API.Experts
 {
+    /// <summary>
+    /// Interface for searching and retrieving expert information.
+    /// </summary>
     public interface IExpertSearcher
     {
-        List<ExpertMetrics> GetExpertMetricsByKeywords(SearchModel searchQuery, TimeRangeModel timeRange);
+        /// <summary>
+        /// Retrieves expert metrics based on the provided search query and time range.
+        /// </summary>
+        /// <param name="searchQuery">The search query.</param>
+        /// <param name="timeRange">The time range.</param>
+        /// <returns>The list of expert metrics.</returns>
+        List<ExpertMetricsDtos> GetExpertMetricsByKeywords(SearchDto searchQuery, TimeRangeDto timeRange);
 
-        ExpertDetails GetExpertDetailsByIdentifier(string identifier);
+        /// <summary>
+        /// Retrieves detailed information about an expert identified by the specified identifier.
+        /// </summary>
+        /// <param name="identifier">The identifier of the expert.</param>
+        /// <returns>The expert details.</returns>
+        ExpertDetailsDto GetExpertDetailsByIdentifier(string identifier);
 
-        List<Collaborator> GetCollaboratorsByExpertIdentifer(string identifer);
+        /// <summary>
+        /// Retrieves a list of collaborators for the expert identified by the specified identifier.
+        /// </summary>
+        /// <param name="identifier">The identifier of the expert.</param>
+        /// <returns>The list of collaborators.</returns>
+        List<CollaboratorDto> GetCollaboratorsByExpertIdentifer(string identifier);
     }
 }
