--- api/Doyen.API/Doyen.API/Models/SearchResultsOrdering.cs
@@ -1,9 +1,0 @@
-﻿namespace Doyen.API.Models
-{
-    public enum SearchResultsOrdering
-    {
-        Relevancy = 0,
-        Publications = 1,
-        Citations = 2
-    }
-}
