--- api/Doyen.API/Doyen.API/Models/TimeRangeModel.cs
@@ -1,9 +1,0 @@
-﻿namespace Doyen.API.Models
-{
-    public class TimeRangeModel
-    {
-        public DateTime GreaterThan { get; set; } = DateTime.Now.AddYears(-5);
-
-        public DateTime LessThan { get; set; } = DateTime.Now;
-    }
-}
